chore: refactored util functions for sheets to a different file (#4378)
