--- analytics/analytics_package/analytics/fields.py
@@ -1,63 +1,0 @@
-# Metric names
-METRIC_EVENT_COUNT = {
-    "id": "eventCount",
-    "alias": "Event Count",
-    "change_alias": "Event Count Change",
-}
-METRIC_TOTAL_USERS = {
-    "id": "totalUsers",
-    "alias": "Total Users",
-    "change_alias": "Total Users Change",
-}
-
-METRIC_ACTIVE_USERS = {
-    "id": "activeUsers",
-    "alias": "Users",
-    "change_alias": "Active Users Change",
-}
-METRIC_PAGE_VIEWS = {
-    "id": "screenPageViews",
-    "alias": "Total Pageviews",
-    "change_alias": "Total Pageviews Change",
-}
-METRIC_SESSIONS = {
-    "id": "sessions",
-    "alias": "Sessions",
-    "change_alias": "Sessions Change",
-}
-SYNTHETIC_METRIC_CLICKS = {
-    "id": None,
-    "alias": "Total Clicks",
-    "change_alias": "Total Clicks Change",
-}
-
-# Event Names
-EVENT_BUILTIN_CLICK = "click"
-EVENT_CUSTOM_CLICK = "outbound_link_clicked"
-EVENT_PAGE_VIEW = "page_view"
-
-# DIMENSIONS
-DIMENSION_PAGE_PATH = {
-    "id": "pagePath",
-    "alias": "Page Path",
-}
-DIMENSION_BUILTIN_URL = {
-    "id": "linkUrl",
-    "alias": "URL",
-}
-DIMENSION_EVENT_NAME = {
-    "id": "eventName",
-    "alias": "Event Name",
-}
-DIMENSION_CUSTOM_URL = {
-    "id": "customEvent:click_url",
-    "alias": "Outbound URL",
-}
-DIMENSION_LANDING_PAGE = {
-    "id": "landingPage",
-    "alias": "Landing Page",
-}
-DIMENSION_YEAR_MONTH = {
-    "id": "yearMonth",
-    "alias": "Month",
-}
